Fix some clang-tidy findings

+++ src/controller/java/AndroidInteractionClient.cpp
@@ -81,7 +81,7 @@ CHIP_ERROR subscribe(JNIEnv * env, jlong handle, jlong callbackHandle, jlong dev
     if (numAttributePaths > 0)
     {
         std::unique_ptr<chip::app::AttributePathParams[]> attributePaths(new chip::app::AttributePathParams[numAttributePaths]);
-        for (uint8_t i = 0; i < numAttributePaths; i++)
+        for (size_t i = 0; i < numAttributePaths; i++)
         {
             jobject attributePathItem = nullptr;
             SuccessOrExit(err = JniReferences::GetInstance().GetListItem(attributePathList, i, attributePathItem));
@@ -107,7 +107,7 @@ CHIP_ERROR subscribe(JNIEnv * env, jlong handle, jlong callbackHandle, jlong dev
     if (numDataVersionFilters > 0)
     {
         std::unique_ptr<chip::app::DataVersionFilter[]> dataVersionFilters(new chip::app::DataVersionFilter[numDataVersionFilters]);
-        for (uint8_t i = 0; i < numDataVersionFilters; i++)
+        for (size_t i = 0; i < numDataVersionFilters; i++)
         {
             jobject dataVersionFilterItem = nullptr;
             SuccessOrExit(err = JniReferences::GetInstance().GetListItem(dataVersionFilterList, i, dataVersionFilterItem));
@@ -141,7 +141,7 @@ CHIP_ERROR subscribe(JNIEnv * env, jlong handle, jlong callbackHandle, jlong dev
     if (numEventPaths > 0)
     {
         std::unique_ptr<chip::app::EventPathParams[]> eventPaths(new chip::app::EventPathParams[numEventPaths]);
-        for (uint8_t i = 0; i < numEventPaths; i++)
+        for (size_t i = 0; i < numEventPaths; i++)
         {
             jobject eventPathItem = nullptr;
             SuccessOrExit(err = JniReferences::GetInstance().GetListItem(eventPathList, i, eventPathItem));
@@ -313,7 +313,7 @@ CHIP_ERROR write(JNIEnv * env, jlong handle, jlong callbackHandle, jlong deviceP
         device->GetExchangeManager(), callback->GetChunkedWriteCallback(),
         convertedTimedRequestTimeoutMs != 0 ? Optional<uint16_t>(convertedTimedRequestTimeoutMs) : Optional<uint16_t>::Missing());
 
-    for (uint8_t i = 0; i < listSize; i++)
+    for (jint i = 0; i < listSize; i++)
     {
         jobject attributeItem             = nullptr;
         jmethodID getEndpointIdMethod     = nullptr;
@@ -494,7 +494,7 @@ CHIP_ERROR extendableInvoke(JNIEnv * env, jlong handle, jlong callbackHandle, jl
     config.SetRemoteMaxPathsPerInvoke(device->GetSecureSession().Value()->GetRemoteSessionParameters().GetMaxPathsPerInvoke());
     SuccessOrExit(err = commandSender->SetCommandSenderConfig(config));
 
-    for (uint8_t i = 0; i < listSize; i++)
+    for (jint i = 0; i < listSize; i++)
     {
         jmethodID getEndpointIdMethod     = nullptr;
         jmethodID getClusterIdMethod      = nullptr;
@@ -845,7 +845,7 @@ CHIP_ERROR ParseAttributePathList(jobject attributePathList, std::vector<app::At
 
     ReturnErrorOnFailure(JniReferences::GetInstance().GetListSize(attributePathList, listSize));
 
-    for (uint8_t i = 0; i < listSize; i++)
+    for (jint i = 0; i < listSize; i++)
     {
         jobject attributePathItem = nullptr;
         ReturnErrorOnFailure(JniReferences::GetInstance().GetListItem(attributePathList, i, attributePathItem));
